requirements.txt
mcp>=1.0.0
playwright>=1.40.0
anthropic>=0.34.0
SpeechRecognition>=3.10.0
PyAudio>=0.2.14
python-dotenv>=1.0.0
